Check for database consistency when obsidian is placed
We previously only checked for removal of portals on BlockBreakEvents.
What some people figured out is that if you can remove the obsidian in
a way that doesn't trigger a BlockBreakEvent, then the database entry
will still be there but there won't be a usable portal.

That way people could create portals that are unusable until somebody
places obsidian back in just the right places.

To stop people from doing this, this method checks for such dangling
database entries whenever somebody places obsidian as well. This
theoretically doesn't remove the problem 100%, if somebody can both
break and then later re-place the obsidian without triggering either
a BlockBreakEvent or a BlockPlaceEvent, then they can still trigger
the old behavior. But that is considered to be so impossible, and
practically impossible for the plugin to monitor for, that we let it be.

In the future we should consider also removing database entries in the
"destination portal frame is missing" case.